[V0 Deprecation] Remove V0 Output Processor (#25320)
